third_party/WebKit/LayoutTests/webaudio/AudioBufferSource/audiobuffersource-detune-modulation-expected.txt

AudioBufferSourceNode: oscillator-driven detune modulation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Exceeded SNR threshold of 93.31 dB.
PASS Maximum difference below threshold of 1.01 ulp (16-bits).
PASS Number of differences between results is 0 out of 11025.
PASS successfullyParsed is true

TEST COMPLETE
